--- FrontEnd/src/pages/User/UserInfoPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserInfoPage = () => {
-  return <div>UserInfoPage</div>;
-};
-
-export default UserInfoPage;
