--- FrontEnd/src/pages/User/UserLoginPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserLoginPage = () => {
-  return <div>UserLoginPage</div>;
-};
-
-export default UserLoginPage;
