--- FrontEnd/src/pages/User/UserSignupPage.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const UserSignupPage = () => {
-  return <div>UserSignupPage</div>;
-};
-
-export default UserSignupPage;
