--- FrontEnd/src/styles/Constants.scss
@@ -1,4 +1,0 @@
-// main color 세팅
-
-$main-color: rgba(255, 255, 255, 0.3);
-$main-white: #eee;
